Fix compile errors on new versions (>8 ?) of GCC.
strncpy -> strcpy is safe, as the destination is allocated with enough space.

The fix for the sprintf is less safe. The error message is (at present)
allocated with 2048 elements, so hopefully this fix is fine as is. Using
snprintf would require knowledge of the error message's size, which could
change, but is not likely and usage of snprintf obscures what the code is trying
to do.