The args struct is now very simple, and is used to fill an obs struct.
gpubox.{c,h} exist to determine things from the gpubox files, and populate the
obs context struct. When populated, this is what will be used to read in the
data.

process_args from args.c is now used to convert the args struct to an obs
struct. The idea is that a simple args struct just contains the metafits and
gpubox files for an observation, and this function just validates and unpacks
those.