Facilitate reading gpubox data.
Examples of usage can be seen in the src/bin and examples directories.

At present, the performance seems pretty good, but perhaps there's a little room
for improvement. The programs that simply sum the values in gpubox files' HDUs
are a good few seconds faster than reading data with mwalib. But maybe that's
because it's much cheaper to continue iterating on the same FITS file pointer
than jumping all over the place.