When inspecting gpubox metadata, open files in parallel.
This is a fairly significant change under the hood. The biggest change is
requiring the fits file handle to be present (i.e. no longer optional) in the
GPUBoxFile struct. To preserve as much of the old code as possible, which allows
us to test other aspects of the gpubox files without them actually existing,
I've introduced a new struct: TempGPUBoxFile struct.

When opening a ~40 GB observation on my HDD with mwalib before this commit,
finding all the UNIX times in the gpubox files takes approximately 30s (!). With
this commit, this is down to 18s. As is documented, the ugly hack I've done to
open the HDUs in parallel will open all fits files twice in total. It's
impractical to do this only once, as sending the FitsFile struct over a thread
has some kind of drawback, and upon opening a file the second time, it should be
virtually instantaneous.

I don't have better ideas for how to reduce the time further. Doing so would
probably compromise how much metadata is presented to a user, and they wouldn't
know how many good UNIX time epochs are available without the existing
functionality.