Fix a bug related to gpubox numbers.
One of the big goals of mwalib is to abstract away obscure details, like the way
that receiver channel numbers behave when bigger than 128. The original
algorithm didn't behave quite right, and is easily fixed with an Option type.

Also made the code a little safer by using a match instead of an if-else, and
not using unwrap. This function should have error handling in the future, though.