Tweak the context display.
I mostly wanted to display the RA and Dec phase centres as normal values (or
"N/A" if applicable) rather than as an Option type. But this commit also
annotates all of the available variables that can be printed, so that things are
clearer.

Also get rid of a bunch of trailing whitespace throughout the file.