Overhaul the error handling in mwalib, and change the API traits.
Previously, traits were tacked onto mwalibContext::new as they made things work,
but we didn't really know what we were doing :) Now that we've seen the light,
removing unneeded traits meant that a bunch of code needed to be tweaked. This
was also a good opportunity to clean up the error handling throughout mwalib.

Interfacing with fits files is now done with a series of macros. This is really
useful as they provide error messages with culprit source files and lines;
previously, it may have been difficult to know which function and which fits
file was misbehaving.

The macros enforce usage of a new FitsError enum. By having this enum separate,
the macros can be used by external crates. A more general MwalibError enum is
introduced, to catch specific oddities.

Removed the .with_context thingy that belongs to the anyhow crate; the new
error-handling code does effectively the same thing. This also means we don't
depend on anyhow in the library.

Simplify some test code changing &str to String; mwalib now no longer cares
between the two.

Fix some documentation grammar.