Expose the fitsio and fitsio-sys crates used.
This allows callers to use whatever version of fitsio and fitsio-sys
that is used by mwalib, in turn ensuring that other dependent libraries
aren't using different versions of these crates. And, along with the
other change to this crate, means that statically-linking cfitsio from
other crates is simpler.

The `infer_static` function introduced to build.rs is a workaround for
pkg-config-rs being too restrictive when static
linking (https://github.com/rust-lang/pkg-config-rs/issues/102).
Basically, if we decide to statically link, we emit a message to cargo,
and it'll work. Hopefully this hack can be removed in the future when
pkg-config-rs is a little more liberal.