Add some keys. Clarify dipole and digital gains.
Add zenith angle, as well as other existing keys but in radians. This
helps the caller by not needing to compute degrees to radians all the
time!

For the RFInput struct, add a dipole_gains member. The type is
convenient for hyperbeam, and will usually only contain ones, except if
that RFInput has a delay of 32 anywhere (code for dead dipole).

As there was already a "gains" member in RFInput, I've changed "gains"
to digital_gains for clarity.

Also stopped all fields from the RFInputMetafitsTableRow struct being
public.