Have cargo use cbindgen to generate a C header.
This avoids a burden on the maintainer to remember to keep the header up
to date, and will automatically re-generate the header whenever the FFI
module changes.

The header is now also C++ compatible, and has a #pragma once line too.

Also delete the C header from the repo. Greg and I discussed this - it's
probably best to use a header that gets generated and not one committed
to the repo.