API tweaks surrounding generic types.
This commit allows multiple generic types on public-facing new methods.
A good example of why this is useful is for the correlator context; when
using a &Path for the metafits, gpubox files must then be a slice of
&Path, which is an unusual collection to have. This commit makes it
possible to use a &Path for the metafits and &[PathBuf] for the
gpuboxes.

Also split the new method logic into new_inner functions. These should
remain private, and this is done to help keep compilation complexity
low. See "Compilation Model: Monomorphization" from this site for more
info: https://matklad.github.io/2021/09/04/fast-rust-builds.html