Speed up legacy order to MWAX order conversion.
Embrace the dark side and use get_unchecked. This eliminates boundary
checks and makes the relevant code about 30% faster. It should also be
safe because the newly-added assert statements ensure that the indices
don't access things they shouldn't. Unfortunately it seems the compiler
isn't smart enough to remove the boundary checks with the asserts alone.