Bump version, MSRV and update coverage CI.
I don't like it, but this
post (https://users.rust-lang.org/t/rust-version-requirement-change-as-semver-breaking-or-not/20980/65)
suggests that it's fine to release a patch version from 0.x.y to
0.x.(y+1) when bumping the MSRV. I wouldn't normally do this, but it's
unlikely to affect anyone, and we're doing it because someone else
bumped MSRV when they probably shouldn't have. /shrug

mwalib's CI was still using the nightly toolchain to generate coverage;
it's now using the stable toolchain.